pyproject.toml
[tool.poetry]
name = "fladar"
version = "0.1.0"
description = "Flight search application to find meeting destinations for two people"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
amadeus = ">=10.0.0"
pyyaml = ">=6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
